# Conflicts:
#	.gitee/ISSUE_TEMPLATE.zh-CN.md
#	ruoyi-common/pom.xml
#	ruoyi-modules/ruoyi-system/pom.xml
#	script/sql/oracle/oracle_ry_job.sql
#	script/sql/postgres/postgres_ry_job.sql
#	script/sql/ry_job.sql
#	smartlink-admin/src/main/java/org/smartlink/web/controller/AuthController.java
#	smartlink-admin/src/main/java/org/smartlink/web/service/SysLoginService.java
#	smartlink-admin/src/main/java/org/smartlink/web/service/impl/EmailAuthStrategy.java
#	smartlink-admin/src/main/java/org/smartlink/web/service/impl/SmsAuthStrategy.java
#	smartlink-admin/src/main/resources/application-prod.yml
#	smartlink-common/smartlink-common-core/src/main/java/org/smartlink/common/core/exception/SseException.java
#	smartlink-common/smartlink-common-core/src/main/java/org/smartlink/common/core/utils/TreeBuildUtils.java
#	smartlink-common/smartlink-common-mail/src/main/java/org/smartlink/common/mail/config/MailConfig.java
#	smartlink-common/smartlink-common-mail/src/main/java/org/smartlink/common/mail/utils/GlobalMailAccount.java
#	smartlink-common/smartlink-common-mail/src/main/java/org/smartlink/common/mail/utils/InternalMailUtil.java
#	smartlink-common/smartlink-common-mail/src/main/java/org/smartlink/common/mail/utils/Mail.java
#	smartlink-common/smartlink-common-mail/src/main/java/org/smartlink/common/mail/utils/MailAccount.java
#	smartlink-common/smartlink-common-mail/src/main/java/org/smartlink/common/mail/utils/MailException.java
#	smartlink-common/smartlink-common-mail/src/main/java/org/smartlink/common/mail/utils/UserPassAuthenticator.java
#	smartlink-common/smartlink-common-redis/src/main/java/org/smartlink/common/redis/utils/CacheUtils.java
#	smartlink-common/smartlink-common-security/src/main/java/org/smartlink/common/security/config/SecurityConfig.java
#	smartlink-common/smartlink-common-tenant/src/main/java/org/smartlink/common/tenant/helper/TenantHelper.java
#	smartlink-common/smartlink-common-tenant/src/main/java/org/smartlink/common/tenant/manager/TenantSpringCacheManager.java
#	smartlink-common/smartlink-common-web/src/main/java/org/smartlink/common/web/config/FilterConfig.java
#	smartlink-common/smartlink-common-web/src/main/java/org/smartlink/common/web/filter/XssFilter.java
#	smartlink-common/smartlink-common-web/src/main/java/org/smartlink/common/web/filter/XssHttpServletRequestWrapper.java
#	smartlink-common/smartlink-common-web/src/main/java/org/smartlink/common/web/handler/GlobalExceptionHandler.java
#	smartlink-common/smartlink-common-web/src/main/java/org/smartlink/common/web/interceptor/PlusWebInvokeTimeInterceptor.java
#	smartlink-extend/smartlink-monitor-admin/src/main/resources/application.yml
#	smartlink-extend/smartlink-snailjob-server/src/main/resources/application-dev.yml
#	smartlink-extend/smartlink-snailjob-server/src/main/resources/application-prod.yml
#	smartlink-modules/smartlink-demo/src/main/java/org/smartlink/demo/mapper/TestDemoMapper.java
#	smartlink-modules/smartlink-demo/src/main/java/org/smartlink/demo/service/impl/TestDemoServiceImpl.java
#	smartlink-modules/smartlink-generator/src/main/java/org/smartlink/generator/config/MyBatisDataSourceMonitor.java
#	smartlink-modules/smartlink-generator/src/main/java/org/smartlink/generator/mapper/GenTableColumnMapper.java
#	smartlink-modules/smartlink-generator/src/main/java/org/smartlink/generator/mapper/GenTableMapper.java
#	smartlink-modules/smartlink-generator/src/main/java/org/smartlink/generator/service/GenTableServiceImpl.java
#	smartlink-modules/smartlink-generator/src/main/resources/mapper/generator/GenTableColumnMapper.xml
#	smartlink-modules/smartlink-generator/src/main/resources/mapper/generator/GenTableMapper.xml
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/controller/monitor/SysLogininforController.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/controller/system/SysDeptController.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/controller/system/SysNoticeController.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/controller/system/SysProfileController.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/controller/system/SysTenantController.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/listener/SysUserImportListener.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/service/ISysTenantPackageService.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/service/impl/SysOperLogServiceImpl.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/service/impl/SysTenantServiceImpl.java
#	smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/service/impl/SysUserServiceImpl.java
#	smartlink-modules/smartlink-workflow/src/main/java/org/smartlink/workflow/service/impl/ActProcessDefinitionServiceImpl.java
#	smartlink-modules/smartlink-workflow/src/main/java/org/smartlink/workflow/service/impl/ActTaskServiceImpl.java
#	smartlink-modules/smartlink-workflow/src/main/java/org/smartlink/workflow/service/impl/WorkflowServiceImpl.java

+++ smartlink-modules/smartlink-system/src/main/java/org/smartlink/system/controller/system/SysDeptController.java
@@ -29,6 +29,7 @@ import java.util.List;
 public class SysDeptController extends BaseController {
 
     private final ISysDeptService deptService;
+    private final ISysPostService postService;
 
     /**
      * 获取部门列表
@@ -116,6 +117,9 @@ public class SysDeptController extends BaseController {
         }
         if (deptService.checkDeptExistUser(deptId)) {
             return R.warn("部门存在用户,不允许删除");
+        }
+        if (postService.countPostByDeptId(deptId) > 0) {
+            return R.warn("部门存在岗位,不允许删除");
         }
         deptService.checkDeptDataScope(deptId);
         return toAjax(deptService.deleteDeptById(deptId));
